move DeveloperAdminComponent to Developers namespace
