--- app/components/developer_admin_component.rb
@@ -1,12 +1,0 @@
-class DeveloperAdminComponent < ApplicationComponent
-  private attr_reader :developer, :user
-
-  def initialize(developer, user:)
-    @developer = developer
-    @user = user
-  end
-
-  def render?
-    !!@user&.admin?
-  end
-end
