+++ app/components/developers/admin_component.html.erb
