+++ app/views/developers/_details.html.erb
@@ -30,6 +30,6 @@
   <% end %>
 
   <%= render RenderableComponent.new("space-y-3 pt-4") do %>
-    <%= render DeveloperAdminComponent.new(developer, user: current_user) %>
+    <%= render Developers::AdminComponent.new(developer, user: current_user) %>
   <% end %>
 </div>
